refactor: 헤더 토큰값 추출 클래스를 auth 패키지로 이동
